loopTextReplace doesn't work, need to revise
For some reason template strings with backticks doesn't activate in the
loopTextReplace for loop, causing the "${x}" to remain the same.

This could either be because of a limitation of template strings, or
some other reason. For now I will push and work on more during the next
week or two to find a solution, as I want the end user to be able to
eventually just do all of this through one function that does callbacks
to all these smaller functions I'm creating.